// src/app/cosmetics/shared/models/string-data-popup-model-with-action.ts
export class StringDataPopupModelWithAction {
    mode: string;
    dataList: any;
    isVisible: boolean;
    pattern: string;
    isSelectable: boolean;
    userData: string;
    columns: Array<{ columnName: string, width: string }>;
    dialogHeight: string;
    dialogWidth: string;
    popupTitle: string;
}